socialapi/requirements.txt
appdirs==1.4.3
astroid==2.2.5
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
clearbit==0.1.7
colorama==0.4.1
cssselect==1.1.0
Django==2.2.6
django-debug-toolbar==2.0
django-filter==2.2.0
django-summernote==0.8.11.4
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
et-xmlfile==1.0.1
fake-useragent==0.1.11
idna==2.8
isort==4.3.21
jdcal==1.4.1
joblib==0.14.0
lazy-object-proxy==1.4.1
lxml==4.4.1
mccabe==0.6.1
numpy==1.17.3
openpyxl==3.0.0
pandas==0.25.3
parse==1.12.1
Pillow==6.2.0
pyee==6.0.0
PyJWT==1.7.1
pylint==2.3.1
pyppeteer==0.0.25
PyQt5==5.13.2
PyQt5-sip==12.7.0
pyquery==1.4.0
python-dateutil==2.8.0
python-slugify==3.0.6
pytz==2019.3
requests==2.22.0
requests-html==0.10.0
scikit-learn==0.21.3
scipy==1.3.2
six==1.12.0
soupsieve==1.9.4
sqlparse==0.3.0
text-unidecode==1.3
tqdm==4.36.1
typed-ast==1.4.0
ultimate-sitemap-parser==0.5
urllib3==1.25.6
w3lib==1.21.0
websockets==8.0.2
wincertstore==0.2
wrapt==1.11.2
xlrd==1.2.0
